fix(trace): satisfy attach audit blockers
